rename the validatedSafeStringFileName
