+++ packages/teams-js/src/public/validatedSafeString.ts
